Change position to float

+++ src/main/java/io/github/jessicacarneiro/apisrest/interfaces/outcoming/output/Position.java
@@ -1,10 +1,9 @@
 package io.github.jessicacarneiro.apisrest.interfaces.outcoming.output;
 
-import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class Position {
-    private BigDecimal lat;
-    private BigDecimal lon;
+    private double lat;
+    private double lon;
 }
